added tensory types

+++ metta/agent/external/example.py
@@ -111,7 +111,8 @@ class Policy(nn.Module):
                 3.0,
                 1.0,
                 2.0,
-            ]
+            ],
+            dtype=torch.float32,
         )[None, :, None, None]
         self.register_buffer("max_vec", max_vec)
